fix path after restructure
