+++ src/frontend/index.html
@@ -4,12 +4,12 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="src/style.css" type="text/css" />
+    <link rel="stylesheet" href="src/app/style.css" type="text/css" />
     <title>Vite + React + TS</title>
     <!--app-head-->
   </head>
   <body>
     <div id="root"><!--app-html--></div>
-    <script type="module" src="/src/entry-client.tsx"></script>
+    <script type="module" src="/src/app/entry-client.tsx"></script>
   </body>
 </html>
