+++ src/frontend/package.json
@@ -8,7 +8,7 @@
         "prod": "NODE_ENV=production node server",
         "build": "npm run build:client && npm run build:server",
         "build:client": "vite build --outDir dist/client",
-        "build:server": "vite build --ssr src/entry-server.tsx --outDir dist/server",
+        "build:server": "vite build --ssr src/app/entry-server.tsx --outDir dist/server",
         "preview": "cross-env NODE_ENV=production node server",
         "test": "vitest",
         "test:run": "vitest run --passWithNoTests",
