+++ src/frontend/server.js
@@ -46,7 +46,7 @@ app.use('*all', async (req, res) => {
       // Always read fresh template in development
       template = await fs.readFile('./index.html', 'utf-8');
       template = await vite.transformIndexHtml(url, template);
-      render = (await vite.ssrLoadModule('/src/entry-server.tsx')).render;
+      render = (await vite.ssrLoadModule('/src/app/entry-server.tsx')).render;
     } else {
       template = templateHtml;
       render = (await import('./dist/server/entry-server.js')).render;
