+++ src/frontend/src/app/App.tsx
@@ -1,17 +1,17 @@
 import { Route, Routes, useLocation } from 'react-router-dom'
 import { useLayoutEffect } from 'react';
 import { Game } from '../pages/Game.tsx'
-import { Menu } from '../pages/Menu.tsx'
-import { StartMenu } from '../pages/StartMenu.tsx'
+import { Home } from '../pages/Home.tsx'
+import { Login } from '../pages/Login.tsx'
 import { AuthProvider } from '../components/providers/Auth.tsx';
 import { RoomProvider } from '../components/providers/Room.tsx';
-import { CONFIG } from '../constants/AppConfig.ts';
 import { ErrorBoundary } from 'react-error-boundary';
-import { GeneralErrorFallback } from '../components/errors/GeneralErrorFallBack.tsx';
 import { Profile } from '../pages/Profile.tsx';
 import { Tournament } from '../pages/Tournament.tsx';
 import { VisitorProfile } from '../pages/VisitorProfile.tsx';
 import { Relationships } from '../pages/Relationships.tsx';
+import { CONFIG } from '../shared/config/AppConfig.ts';
+import { GeneralErrorFallback } from '../features/errors/GeneralErrorFallBack.tsx';
 
 /* v8 ignore start */
 export function App() {
@@ -33,8 +33,8 @@ export function App() {
 
                 <RoomProvider>
                     <Routes>
-                        <Route path={CONFIG.START_MENU_NAVIGATION} element={<StartMenu />} />
-                        <Route path={CONFIG.MENU_NAVIGATION} element={<Menu />} />
+                        <Route path={CONFIG.START_MENU_NAVIGATION} element={<Login />} />
+                        <Route path={CONFIG.MENU_NAVIGATION} element={<Home />} />
                         <Route path={CONFIG.USER_PROFILE_NAVIGATION} element={<Profile />} />
                         <Route path={CONFIG.USER_RELATIONSHIPS_NAVIGATION} element={<Relationships />} />
                         <Route path={CONFIG.VISITOR_PROFILE_NAVIGATION} element={<VisitorProfile />} />
