+++ src/frontend/src/components/layout/CenteredFlexContainer.tsx
@@ -1,0 +1,11 @@
+import { JSX, ReactNode } from "react";
+
+export function CenterFlexContainer({ children }: { children: ReactNode }): JSX.Element {
+    return (
+        <div className="flex justify-between min-h-full grow">
+            <div />
+            {children}
+            <div />
+        </div>
+    )
+}
