+++ src/frontend/src/components/layout/MainWindowContainer.tsx
@@ -1,0 +1,10 @@
+import { ReactNode } from "react";
+import { JSX } from "react/jsx-runtime";
+
+export function MainWindowContainer({ children }: { children: ReactNode }): JSX.Element {
+    return (
+        <div id='main-window-container' className="min-w-full min-h-full flex flex-col">
+            {children}
+        </div>
+    )
+}
