+++ src/frontend/src/features/login/Title.tsx
@@ -1,0 +1,5 @@
+import { JSX } from "react";
+
+export function Title(): JSX.Element {
+    return (<h1 className="text-8xl font-monof text-metalgear">Counter-pong</h1>)
+}
