+++ src/frontend/src/features/neofetch/Neofetch.tsx
@@ -1,6 +1,5 @@
 import { JSX } from 'react';
 import platform from 'platform';
-import { browsers } from "../utils/browserLogos";
 
 /* v8 ignore start*/
 export function Neofetch(): JSX.Element {
@@ -38,4 +37,209 @@ export function Neofetch(): JSX.Element {
         </div>
     )
 }
+
+export const browsers = new Map([
+    [
+        'Firefox',
+        {
+            name: 'Firefox',
+            displayName: 'Mozilla Firefox',
+            logo: `                               .:                 
+                             :::::.               
+                           .::::::::              
+                          :::::::::::.            
+        ::               .:::::::::::::.          
+      =*+=    :-.        ::::::::::::::::         
+    -###*+ .+#*+.       :---::::::::::::::. :.    
+   *%%###*+*****:  :-=+*****++=-:::::::::::: ::   
+  *%%%%%###***+++=**************+=-:::::::::::::  
+ +%%%%%%%####***+++++*********+===+=::::::::::::: 
+.%%%%%%#**++++++**+++++++*******+-:--------::::::.
++%%%%%%*+++====---------+*********=:::-=------::::
+#%%%%%%%%#**+=====+++++************----======-::::
+%%%%%%%%%%%%*=+*###***************#+---=++++-::---
+#%%%%%%%%%%%#+###########********##+====#**=---==-
++%%%%%%%%%%%%*#%%###########***####++***%#===-=++-
+.%%%%%%%%%%%%%##%%%%####################*+++==***.
+ =%%%%%%%%%%%%%%##%%%%%%%%######%%%%%%#*****+###= 
+  *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%####%%%*  
+   *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*   
+    =%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%=    
+     .*%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#.     
+       :*%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*:       
+         .=#%%%%%%%%%%%%%%%%%%%%%%%%%%#=.         
+            .-+#%%%%%%%%%%%%%%%%%%#*-.            
+`
+        }
+    ],
+    [
+        'Chrome',
+        {
+            name: 'Chrome',
+            displayName: 'Google Chrome',
+            logo: `                .-=+*##%%%%%#*+=-.                
+            :=*%%%%%%%%%%%%%%%%%%%%*=:            
+         :+%%%%%%%%%%%%%%%%%%%%%%%%%%%%+.         
+       -#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:       
+     :#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:     
+    =%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%=    
+   -+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*   
+  -===*%%%%%%%%%%%%*=:.     ..:----------------:  
+ :=====*%%%%%%%%%=  ..::::::..  .:::::::::::::::: 
+ =======+%%%%%%+  :------------:  .:::::::::::::: 
+:========+%%%%+ .================  :::::::::::::::
+==========+#%%  +++++++++++++++++=  ::::::::::::::
+===========+## .******************  ::::::::::::::
+=============* .******************  ::::::::::::::
+-=============: -###############*: .:::::::::::::.
+.==============: :*############*. .:::::::::::::: 
+ -==============-. :+*######*=: .:::::::::::::::. 
+  =================:.  ....  .-+-:::::::::::::::  
+   =======================+***+-:::::::::::::::   
+    -====================+***+-:::::::::::::::    
+     .===================+**=:::::::::::::::.     
+       :================+**=::::::::::::::.       
+         .-=============+*-:::::::::::::          
+            .-=========++-::::::::::.             
+                .:--====-::::::..                 
+`
+        }
+    ],
+    [
+        'chromium',
+        {
+            name: 'Chromium',
+            displayName: 'Chromium',
+            logo: `                .-=+*##%%%%%#*+=-.                
+            :=*%%%%%%%%%%%%%%%%%%%%*=:            
+         :+%%%%%%%%%%%%%%%%%%%%%%%%%%%%+.         
+       -#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:       
+     :#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:     
+    =%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%=    
+   -+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*   
+  -===*%%%%%%%%%%%%*=:.     ..:----------------:  
+ :=====*%%%%%%%%%=  ..::::::..  .:::::::::::::::: 
+ =======+%%%%%%+  :------------:  .:::::::::::::: 
+:========+%%%%+ .================  :::::::::::::::
+==========+#%%  +++++++++++++++++=  ::::::::::::::
+===========+## .******************  ::::::::::::::
+=============* .******************  ::::::::::::::
+-=============: -###############*: .:::::::::::::.
+.==============: :*############*. .:::::::::::::: 
+ -==============-. :+*######*=: .:::::::::::::::. 
+  =================:.  ....  .-+-:::::::::::::::  
+   =======================+***+-:::::::::::::::   
+    -====================+***+-:::::::::::::::    
+     .===================+**=:::::::::::::::.     
+       :================+**=::::::::::::::.       
+         .-=============+*-:::::::::::::          
+            .-=========++-::::::::::.             
+                .:--====-::::::..                 
+`
+        }
+    ],
+    [
+        'brave',
+        {
+            name: 'Brave',
+            displayName: 'Brave',
+            logo: `              .+##################+.              
+             =*#####################=             
+      .... :*#########################-..::.      
+   .=*******############################%%%%%+:   
+ .=*********+**######*+=--=+*########*%%%%%%%%%*. 
+:**********:    ..:.          .:..    :#%%%%%%%%%:
+ =*******=                              =%%%%%%%+ 
+ -******.   .:::::.           .:---::.   .#%%%%%= 
+.*****=      :-=+****+:    :*#%%%#+=:      +%%%%%:
++****+            .-*#.    .%#=:            *%%%%*
+-*****:             =+      **             :#%%%%-
+ ******+:          -*-      -%=          :*%%%%%# 
+ :*******+.       =**        %%+       :*%%%%%%%- 
+  +********+.     =**-:...::=%#+     .+%%%%%%%%#  
+  :**********.      :=*#####+:      .%%%%%%%%%%-  
+   +********+          :**:          *%%%%%%%%*   
+   :********:          :##-          :%%%%%%%%:   
+    +*******+        -*#####=.       *%%%%%%%*    
+    .*********-..:=+**=-::-+#%#+-..-#%%%%%%%%:    
+     =************=:         .-+#%%%%%%%%%%%+     
+     .***********-              =%%%%%%%%%%%.     
+      =************+:        :+%%%%%%%%%%%%+      
+       +**************=.  .=#%%%%%%%%%%%%%#       
+        =************###++##%%%%%%%%%%%%%*        
+         .=**********#######%%%%%%%%%%#+:         
+            :=*******#######%%%%%%%%+-            
+               :+****#######%%%%%*-               
+                  :+*#######%%*-                  
+                     :+####*-                     
+`
+        }
+    ],
+    [
+        'edge',
+        {
+            name: 'Edge',
+            displayName: 'Microsoft Edge',
+            logo: `                 ...:::::::::::...                
+             .:::::::::::::::::-----:.            
+          .::::::::::::::::::::::-------:.        
+        :::::::::::::::::::::::::::-------:.      
+      ::::::::::::::::::::::::::::::--------:.    
+    .:::::::::::::::::::::::::::::::::------::.   
+   .:::--===++++=++===--:::::::::::::-----::::::  
+  ::-==++=================-::::::---------::::::: 
+ .-=++======================-::-------------:::::.
+ -++============------:...::==-------------------:
+:+==================.        .--------------------
+=================+*            ------------------:
+================**=            :-----------------.
+-==============***+            -----------------. 
+:=============*****:          -----------------   
+ ============+*****+:        .---==========-:.    
+ :===========********-          ..:::::::.        
+  -==========+********+-                          
+   -=========+***********=-.                .:    
+    :=========***************+=-::....::-=+**=    
+     .-========****************************+.     
+       .-=======+************************+:       
+         .-======+*********************=.         
+            .:-====++**************+-:            
+                ..:--==++++++==-:.                
+`
+        }
+    ],
+    [
+        'opera',
+        {
+            name: 'Opera',
+            displayName: 'Opera',
+            logo: `                .:-=++******++=-:.                
+            :-+**************#####*+-:            
+         .=************##%%%%%%%%%%%%%#+:         
+       :+************+=-:::-=*%%%%%%%%%%%#-       
+     :+***********-.           -*##########*:     
+    =***********=                :*##########=    
+   +***********:                   =##########+   
+  +***********.                     =**********+  
+ =***********:                       +**********= 
+.***********+                        .***********.
+=***********:                         +++++++++++-
++***********.                         -++++++++++=
+************                          :+++++++++++
++***********.                         -===========
+=***********:                         -==========:
+.***********+                        .===========.
+ =***********:                       -----------: 
+  +***********.                     :-----------  
+   +##########*:                   :----------:   
+    =###########=                .-----------:    
+     :*###########-            .:-----------.     
+       -*###########+-......::------------.       
+         :+############*+=-----------==-.         
+            :=*%%%%%%%%%%%%##*****#*=:            
+                :-=+*##%%%%##*+=-:                
+`
+        }
+    ]
+]);
 /* v8 ignore stop*/
