+++ src/frontend/src/features/profile/DeleteAccount.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { useAuth } from "../../components/providers/Auth";
-import { getCookie } from "../../components/utils/cookies";
-import api from "../../api";
-import { CONFIG } from "../../constants/AppConfig";
+import api from "../../shared/api/api";
+import { getCookie } from "../../shared/utils/cookies";
+import { CONFIG } from "../../shared/config/AppConfig";
 import { SubmitPropertyChangeYesNo } from "./Submit";
