+++ src/frontend/src/features/profile/ProfileRelationships.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom"
-import { CONFIG } from "../../constants/AppConfig";
+import { CONFIG } from "../../shared/config/AppConfig";
 
 export function ProfileRelationships() {
     const navigate = useNavigate();
