+++ src/frontend/src/features/relationships/FriendshipRequests.tsx
@@ -1,9 +1,9 @@
 import { JSX } from "react";
-import { IFriendship } from "../../types/friendship";
-import api from "../../api";
-import { CONFIG } from "../../constants/AppConfig";
-import { Terminal } from "../../components/utils/MenuUtils";
+import api from "../../shared/api/api";
+import { CONFIG } from "../../shared/config/AppConfig";
 import { RelationsColumn } from "./RelationsColumn";
+import { Terminal } from "../../components/layout/Terminal";
+import { IFriendship } from "../../shared/types/friendship";
 
 export function FriendshipRequests({ requests }: { requests: IFriendship[] }) {
