+++ src/frontend/src/features/relationships/RelationsColumn.tsx
@@ -1,4 +1,3 @@
-
 import { ReactNode } from "react";
 
 export function RelationsColumn({ children }: { children: ReactNode }) {
