--- src/frontend/src/shared/utils/browserLogos.ts
@@ -1,204 +1,0 @@
-export const browsers = new Map([
-  [
-    'Firefox',
-    {
-      name: 'Firefox',
-      displayName: 'Mozilla Firefox',
-      logo: `                               .:                 
-                             :::::.               
-                           .::::::::              
-                          :::::::::::.            
-        ::               .:::::::::::::.          
-      =*+=    :-.        ::::::::::::::::         
-    -###*+ .+#*+.       :---::::::::::::::. :.    
-   *%%###*+*****:  :-=+*****++=-:::::::::::: ::   
-  *%%%%%###***+++=**************+=-:::::::::::::  
- +%%%%%%%####***+++++*********+===+=::::::::::::: 
-.%%%%%%#**++++++**+++++++*******+-:--------::::::.
-+%%%%%%*+++====---------+*********=:::-=------::::
-#%%%%%%%%#**+=====+++++************----======-::::
-%%%%%%%%%%%%*=+*###***************#+---=++++-::---
-#%%%%%%%%%%%#+###########********##+====#**=---==-
-+%%%%%%%%%%%%*#%%###########***####++***%#===-=++-
-.%%%%%%%%%%%%%##%%%%####################*+++==***.
- =%%%%%%%%%%%%%%##%%%%%%%%######%%%%%%#*****+###= 
-  *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%####%%%*  
-   *%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*   
-    =%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%=    
-     .*%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#.     
-       :*%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*:       
-         .=#%%%%%%%%%%%%%%%%%%%%%%%%%%#=.         
-            .-+#%%%%%%%%%%%%%%%%%%#*-.            
-`
-    }
-  ],
-  [
-    'Chrome',
-    {
-      name: 'Chrome',
-      displayName: 'Google Chrome',
-      logo: `                .-=+*##%%%%%#*+=-.                
-            :=*%%%%%%%%%%%%%%%%%%%%*=:            
-         :+%%%%%%%%%%%%%%%%%%%%%%%%%%%%+.         
-       -#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:       
-     :#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:     
-    =%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%=    
-   -+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*   
-  -===*%%%%%%%%%%%%*=:.     ..:----------------:  
- :=====*%%%%%%%%%=  ..::::::..  .:::::::::::::::: 
- =======+%%%%%%+  :------------:  .:::::::::::::: 
-:========+%%%%+ .================  :::::::::::::::
-==========+#%%  +++++++++++++++++=  ::::::::::::::
-===========+## .******************  ::::::::::::::
-=============* .******************  ::::::::::::::
--=============: -###############*: .:::::::::::::.
-.==============: :*############*. .:::::::::::::: 
- -==============-. :+*######*=: .:::::::::::::::. 
-  =================:.  ....  .-+-:::::::::::::::  
-   =======================+***+-:::::::::::::::   
-    -====================+***+-:::::::::::::::    
-     .===================+**=:::::::::::::::.     
-       :================+**=::::::::::::::.       
-         .-=============+*-:::::::::::::          
-            .-=========++-::::::::::.             
-                .:--====-::::::..                 
-`
-    }
-  ],
-  [
-    'chromium',
-    {
-      name: 'Chromium',
-      displayName: 'Chromium',
-      logo: `                .-=+*##%%%%%#*+=-.                
-            :=*%%%%%%%%%%%%%%%%%%%%*=:            
-         :+%%%%%%%%%%%%%%%%%%%%%%%%%%%%+.         
-       -#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:       
-     :#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%#:     
-    =%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%=    
-   -+#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%*   
-  -===*%%%%%%%%%%%%*=:.     ..:----------------:  
- :=====*%%%%%%%%%=  ..::::::..  .:::::::::::::::: 
- =======+%%%%%%+  :------------:  .:::::::::::::: 
-:========+%%%%+ .================  :::::::::::::::
-==========+#%%  +++++++++++++++++=  ::::::::::::::
-===========+## .******************  ::::::::::::::
-=============* .******************  ::::::::::::::
--=============: -###############*: .:::::::::::::.
-.==============: :*############*. .:::::::::::::: 
- -==============-. :+*######*=: .:::::::::::::::. 
-  =================:.  ....  .-+-:::::::::::::::  
-   =======================+***+-:::::::::::::::   
-    -====================+***+-:::::::::::::::    
-     .===================+**=:::::::::::::::.     
-       :================+**=::::::::::::::.       
-         .-=============+*-:::::::::::::          
-            .-=========++-::::::::::.             
-                .:--====-::::::..                 
-`
-    }
-  ],
-  [
-    'brave',
-    {
-      name: 'Brave',
-      displayName: 'Brave',
-      logo: `              .+##################+.              
-             =*#####################=             
-      .... :*#########################-..::.      
-   .=*******############################%%%%%+:   
- .=*********+**######*+=--=+*########*%%%%%%%%%*. 
-:**********:    ..:.          .:..    :#%%%%%%%%%:
- =*******=                              =%%%%%%%+ 
- -******.   .:::::.           .:---::.   .#%%%%%= 
-.*****=      :-=+****+:    :*#%%%#+=:      +%%%%%:
-+****+            .-*#.    .%#=:            *%%%%*
--*****:             =+      **             :#%%%%-
- ******+:          -*-      -%=          :*%%%%%# 
- :*******+.       =**        %%+       :*%%%%%%%- 
-  +********+.     =**-:...::=%#+     .+%%%%%%%%#  
-  :**********.      :=*#####+:      .%%%%%%%%%%-  
-   +********+          :**:          *%%%%%%%%*   
-   :********:          :##-          :%%%%%%%%:   
-    +*******+        -*#####=.       *%%%%%%%*    
-    .*********-..:=+**=-::-+#%#+-..-#%%%%%%%%:    
-     =************=:         .-+#%%%%%%%%%%%+     
-     .***********-              =%%%%%%%%%%%.     
-      =************+:        :+%%%%%%%%%%%%+      
-       +**************=.  .=#%%%%%%%%%%%%%#       
-        =************###++##%%%%%%%%%%%%%*        
-         .=**********#######%%%%%%%%%%#+:         
-            :=*******#######%%%%%%%%+-            
-               :+****#######%%%%%*-               
-                  :+*#######%%*-                  
-                     :+####*-                     
-`
-    }
-  ],
-  [
-    'edge',
-    {
-      name: 'Edge',
-      displayName: 'Microsoft Edge',
-      logo: `                 ...:::::::::::...                
-             .:::::::::::::::::-----:.            
-          .::::::::::::::::::::::-------:.        
-        :::::::::::::::::::::::::::-------:.      
-      ::::::::::::::::::::::::::::::--------:.    
-    .:::::::::::::::::::::::::::::::::------::.   
-   .:::--===++++=++===--:::::::::::::-----::::::  
-  ::-==++=================-::::::---------::::::: 
- .-=++======================-::-------------:::::.
- -++============------:...::==-------------------:
-:+==================.        .--------------------
-=================+*            ------------------:
-================**=            :-----------------.
--==============***+            -----------------. 
-:=============*****:          -----------------   
- ============+*****+:        .---==========-:.    
- :===========********-          ..:::::::.        
-  -==========+********+-                          
-   -=========+***********=-.                .:    
-    :=========***************+=-::....::-=+**=    
-     .-========****************************+.     
-       .-=======+************************+:       
-         .-======+*********************=.         
-            .:-====++**************+-:            
-                ..:--==++++++==-:.                
-`
-    }
-  ],
-  [
-    'opera',
-    {
-      name: 'Opera',
-      displayName: 'Opera',
-      logo: `                .:-=++******++=-:.                
-            :-+**************#####*+-:            
-         .=************##%%%%%%%%%%%%%#+:         
-       :+************+=-:::-=*%%%%%%%%%%%#-       
-     :+***********-.           -*##########*:     
-    =***********=                :*##########=    
-   +***********:                   =##########+   
-  +***********.                     =**********+  
- =***********:                       +**********= 
-.***********+                        .***********.
-=***********:                         +++++++++++-
-+***********.                         -++++++++++=
-************                          :+++++++++++
-+***********.                         -===========
-=***********:                         -==========:
-.***********+                        .===========.
- =***********:                       -----------: 
-  +***********.                     :-----------  
-   +##########*:                   :----------:   
-    =###########=                .-----------:    
-     :*###########-            .:-----------.     
-       -*###########+-......::------------.       
-         :+############*+=-----------==-.         
-            :=*%%%%%%%%%%%%##*****#*=:            
-                :-=+*##%%%%##*+=-:                
-`
-    }
-  ]
-]);
