+++ src/frontend/src/shared/utils/html.ts
@@ -1,0 +1,3 @@
+export function htmlIdefier(id: string) {
+  return id.toLowerCase().trim().replace(' ', '-');
+}
